Rename StarCraftEnv and RemoteEnvAPIClient
