+++ gym/envs/starcraft/starcraft_basic_env.py
@@ -1,5 +1,4 @@
-from gym.envs.StarCraftEnv import StarCraftEnv
-
+from gym.envs.starcraft.starcraft_env import StarCraftEnv
 
 class StarCraftBasicEnv(StarCraftEnv):
     def __init__(self):
