+++ gym/envs/starcraft/starcraft_image_file.py
@@ -19,6 +19,7 @@ class StarCraftImageFile(ImageFile.ImageFile):
 
 
     # StarCraft uses a fixed 256 color palette. We keep the palette here for fast decoding.
+    # If this class ends up being slow, consider ditching the PIL implementation and just use np
     _palette_bytes = b'\x00\x00\x00##\xff##\xff##\xff##\xff##\xff##\xff##\xff\xff\x00\xff\xde\x00\xde\xbd\x00\xbd\x9c\x00\x9c|\x00|[\x00[:\x00:\x19\x00\x19,$\x18H$\x14\\,\x14p0\x14h<$|@\x18xL,\xa8\x08\x08\x8cT0\x84`D\xa0T\x1c\xc4L\x18\xbch$\xb4p<\xd0d \xdc\x944\xe0\x94T\xec\xc4T4D(@l<HlPL\x80PP\x8c\\\\\xa0x\x00\x00\x18\x00\x104\x00\x08P$4H0@T\x144|4Ll@XtHh\x8c\x00p\x9cX\x80\xa4@h\xd4\x18\xac\xb8$$\xfcd\x94\xbcp\xa8\xcc\x8c\xc0\xd8\x94\xdc\xf4\xac\xdc\xe8\xac\xfc\xfc\xcc\xf8\xf8\xfc\xfc\x00\xf4\xe4\x90\xfc\xfc\xc0\x0c\x0c\x0c\x18\x14\x10\x1c\x1c ((080$8<DL@0LLL\\P@XXXhhhx\x84lh\x94lt\xa4|\x98\x94\x8c\x90\xb8\x94\x98\xc4\xa8\xb0\xb0\xb0\xac\xcc\xb0\xc4\xc0\xbc\xcc\xe0\xd0\xf0\xf0\xf0\x1c\x10\x08(\x18\x0c4\x10\x084 \x0c8\x10 4( D4\x08H0\x18`\x00\x00T( P@\x14\\T\x14\x84\x04\x04hL4|80pd |PP\xa44\x1c\x94l\x00\x98\\@\x8c\x804\x98tT\xb8TD\xb0\x90\x18\xb0t\\\xf4\x04\x04\xc8xT\xfchT\xe0\xa4\x84\xfc\x94h\xfc\xcc,\x10\xfc\x18\x0c\x00 \x1c\x1c,$$L(,h,0\x84 \x18\xb84<\xachh\x94d\x90\xfc|\xac\xfc\x00\xe4\xfc\x9c\x90@\xa8\x94T\xbc\xa4\\\xcc\xb8`\xe8\xd8\x80\xec\xc4\xb0\xfc\xfc8\xfc\xfc|\xfc\xfc\xa4\x08\x08\x08\x10\x10\x10\x18\x18\x18(((444L<8DDDHHXXXhth8xd\\``|\x84tt\x84\x84\x9c\xac\x8c|\xac\x98\x94\x90\x90\xb8\xb8\xb8\xe8\xf8\x8c\x14\x10T< \x90p,\xb4\x94\x04 dH\x1cP\x084\x98h0x\x88@\x9c\x0cH\xcc\xbc\xb84\xdc\xdc<\x10\x00\x00$\x00\x004\x00\x00H\x00\x00`\x18\x04\x8c(\x08\xc8\x18\x18\xe0,,\xe8  \xe8P\x14\xfc  \xe8x$\xf8\xac<\x00\x14\x00\x00(\x00\x00D\x00\x00d\x00\x08\x80\x08$\x98$<\x9c<X\xb0Xh\xb8h\x80\xc4\x80\x94\xd4\x94\x0c\x14$$<d0P\x848\\\x94Ht\xb4T\x84\xc4`\x94\xd4x\xb4\xec\x04\x04\x04\x14\x14\x14$ \x1c  $$$(,,,,0404<48@<<<<@H@DLLPX```tpp||| \x18\x10<,\x18\x10\x10\x0c\x14\x14\x18\x18\x18\x1c \x1c\x18 \x18\x1c\x1c $$\x1c ( $$(,0$,(,4000,08<,4888D8,@0<D4@P<HHHHDHPHLTTPPPT\\dddpllxtt\x88\x84\x84##\xff##\xff##\xff##\xff##\xff##\xff##\xff##\xff##\xff\xff\xff\xff'
 
     global_palette = ImagePalette.raw("RGB", _palette_bytes)
